+++ src/coreclr/nativeaot/System.Private.CoreLib/src/System/Reflection/Assembly.NativeAot.cs
@@ -2,8 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Configuration.Assemblies;
+using System.Diagnostics;
 using System.IO;
 using System.Runtime.Serialization;
+using System.Security;
 
 using Internal.Reflection.Augments;
 
@@ -16,12 +18,42 @@ namespace System.Reflection
         [System.Runtime.CompilerServices.Intrinsic]
         public static Assembly GetExecutingAssembly() { throw NotImplemented.ByDesign; } //Implemented by toolchain.
 
+        [DynamicSecurityMethod]
         public static Assembly GetCallingAssembly()
         {
             if (AppContext.TryGetSwitch("Switch.System.Reflection.Assembly.SimulatedCallingAssembly", out bool isSimulated) && isSimulated)
                 return GetEntryAssembly();
 
-            throw new PlatformNotSupportedException();
+            if (!StackTrace.IsSupported)
+                throw new NotSupportedException(SR.NotSupported_StackTraceSupportDisabled);
+
+            // We want to be able to handle GetCallingAssembly being called from Main (CoreCLR returns
+            // the assembly of Main), and also GetCallingAssembly being called from things like
+            // delegate invoke thunks. We do this by making the the definition of "calling assembly"
+            // a bit more loose.
+
+            // Technically we want skipFrames: 2 since we're interested in the frame that
+            // called the method that calls GetCallingAssembly, but we might need the first frame
+            // later in this method.
+            var stackTrace = new StackTrace(skipFrames: 1);
+
+            DiagnosticMethodInfo? dmi = null;
+
+            // Note: starting at index 1 since we want to skip the method that called GetCallingAssembly.
+            // We do a foreach so that we can skip any compiler-generated thunks that don't have method info.
+            for (int i = 1; i < stackTrace.FrameCount; i++)
+            {
+                dmi = DiagnosticMethodInfo.Create(stackTrace.GetFrame(i));
+                if (dmi != null)
+                    break;
+            }
+
+            // If we haven't found anything in the entire stack trace, fall back
+            // to the method that called this method. This simulates what CoreCLR would
+            // do if GetCallingAssembly is called from e.g. Main.
+            dmi ??= stackTrace.GetFrame(0) is StackFrame sf ? DiagnosticMethodInfo.Create(sf) : null;
+
+            return dmi.DeclaringAssemblyName is string asmName ? Load(asmName) : null;
         }
 
         public static Assembly Load(AssemblyName assemblyRef) => ReflectionAugments.ReflectionCoreCallbacks.Load(assemblyRef, throwOnFileNotFound: true);
